requirements.txt
pandas==2.2.3
numpy==2.2.4
customtkinter==5.2.2
pillow==11.1.0